[test] Fix warnings in tests on 64-bit systems
Cast argiments from size_t to int.  size_t is 64-bit on 64-bit systems,
which causes a warning.  The actual data should fit 32 bit comfortably.

+++ test/cairo-test.c
@@ -1890,7 +1890,7 @@ cairo_test_expecting (cairo_test_t *test,
 	    }
 
 	    if (!found) {
-		fprintf (stderr, "Cannot test target '%.*s'\n", end - tname, tname);
+		fprintf (stderr, "Cannot test target '%.*s'\n", (int)(end - tname), tname);
 		exit(-1);
 	    }
