+++ test/trap-clip.c
@@ -191,7 +191,7 @@ draw (cairo_t *cr, int width, int height)
 		pattern_funcs[i] (cr, x, y);
 		draw_funcs[j] (cr, x, y);
 		if (cairo_status (cr))
-		    cairo_test_log ("%d %d HERE!\n", i, j);
+		    cairo_test_log ("%d %d HERE!\n", (int)i, (int)j);
 
 		cairo_restore (cr);
 	    }
@@ -199,7 +199,7 @@ draw (cairo_t *cr, int width, int height)
     }
 
     if (cairo_status (cr) != CAIRO_STATUS_SUCCESS)
-	cairo_test_log ("%d %d .HERE!\n", i, j);
+	cairo_test_log ("%d %d .HERE!\n", (int)i, (int)j);
 
     return CAIRO_TEST_SUCCESS;
 }
